[Modify]修改getAllPostAPI
